tests gmock
