+++ tgbot_lib/CMakeLists.txt
@@ -1,0 +1,14 @@
+cmake_minimum_required(VERSION 3.0.0)
+project(tgbot_lib)
+
+set(CMAKE_CXX_STANDARD 17)
+set(TGBOT_LIB_LIBRARIES)
+set(TGBOT_LIB_INCLUDE_DIRS)
+
+file(GLOB SOURCES *.cpp)
+file(GLOB INCLUDE_DIRS ${CMAKE_CURRENT_SOURCE_DIR})
+
+add_library(${PROJECT_NAME} STATIC ${SOURCES})
+
+set(TGBOT_LIB_LIBRARIES ${PROJECT_NAME} PARENT_SCOPE)
+set(TGBOT_LIB_INCLUDE_DIRS ${INCLUDE_DIRS} PARENT_SCOPE)
